+++ CHANGELOG.md
@@ -1,7 +1,37 @@
 # Change Log
-All notable changes to the "multi-command" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## 1.6.0
+### Added
+- [Variable substitution like `${userHome}` or `${config:editor.fontSize}` in arguments.](/README.md#pass-arguments-to-commands)
+- [Conditioned commands to branch a sequence when a command fails.](/README.md#conditioned-commands)
+- [Parameter to repeat a command.](/README.md#repeat-commands)
+- [Language filter for manual execution.](/README.md#manual-execution)
 
-## [Unreleased]
-- Initial release
+## 1.5.1
+### Added
+- Support extensionKind ui or workspace.
+
+## 1.5.0
+### Added
+- Simple usage only with keybindings.json  
+- Object style settings for merging user settings and workspace settings.
+
+## 1.4.0
+### Added
+- New style for binding a key to created commands.
+
+## 1.3.0
+### Added
+- Manual execution from command palette.
+
+## 1.2.0
+### Added
+- Pass arguments to commands.
+
+## 1.1.0
+### Fixed
+- Reloads settings.json when the file is changed.  
+  Now, you can use a custom multi-command immediately after adding it in the settings.json without restarting vscode.
+
+## 1.0.0
+Initial release.